requirements.txt
#jupyter_client
#jupyter_core
#matplotlib
#matplotlib-inline
NavPy==1.0
numpy==1.24.3
numpy-stl==3.1.1
scipy
pyserial
